[Internal] Rename resources directory to products in pluginframework
