+++ internal/acceptance/pipeline_test.go
@@ -88,6 +88,8 @@ func TestAccPipelineResource_CreatePipeline(t *testing.T) {
 				}
 			}
 
+			schema = null
+
 			continuous = false
 		}
 		` + dltNotebookResource,
