+++ internal/providers/pluginfw/products/library/resource_library.go
